requirements.txt
numpy
matplotlib
argparse
tqdm
pyyaml
scipy
scikit-image>=0.13